requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
reportlab
python-multipart